Utils canadapost

+++ shipverse/upsviews.py
@@ -230,14 +230,17 @@ def removeUser(request):
     user = Users.objects.get(id=user_id)
     if not user : 
         return JsonResponse({"message": "User not found or Unauthorized !"},status=status.HTTP_200_OK)
-    shipperno = user_data['shipperno']
+    id = user_data['shipperno'].get("id")
+
     try:
-        shipperLocation = UserCarrier.objects.get(
-            user=user, account_number=shipperno)   # Updated accountNumber to account_number
+        shipperLocation = UserCarrier.objects.get(id = id) 
+        if not shipperLocation:
+            return JsonResponse({'success': False}, status=status.HTTP_404_NOT_FOUND)
+        shipperLocation.delete()
+        return JsonResponse({'success': True}, status=status.HTTP_200_OK)
     except:
         return JsonResponse({'success': False}, status=status.HTTP_404_NOT_FOUND)
-    shipperLocation.delete()
-    return JsonResponse({'success': True}, status=status.HTTP_200_OK)
+    
 
 
 @api_view(['POST'])
